Testing a shell script that will run the preprocessor
I started by compiling the JavaCC tokenizer in the last push. In this push I'm testing checking if a data directory exists in the root script directory. If the data directory does exist, I'm deleting it. After that control structure I'm creating a new data directory. This is where I'll direct the output of the tokenizer.